requirements.txt
annotated-types==0.7.0
anyio==3.7.1
asgiref==3.8.1
beautifulsoup4==4.13.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
Django==5.1.1
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
filelock==3.15.4
fonttools==4.56.0
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
kiwisolver==1.4.8
lxml==5.3.1
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.3
numpy==2.0.1
openai==1.0.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==73.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
